ANDP/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.7.5
annotated-types==0.6.0
ansible==5.9.0
ansible-core==2.12.10
anyio==4.3.0
argcomplete==2.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
asttokens==3.0.0
async-timeout==4.0.2
attrs==20.3.0
Authlib==1.1.0
azure-common==1.1.22
azure-core==1.25.0
azure-identity==1.10.0
azure-mgmt-core==1.3.1
azure-mgmt-rdbms==10.2.0b1
azure-mgmt-resource==21.1.0b1
azure-mgmt-subscription==3.0.0
Babel==2.9.1
bcrypt==3.2.2
Beaker==1.10.0
beautifulsoup4==4.12.2
bidict==0.21.4
black==25.1.0
bleach==6.2.0
blinker==1.4
blivet==3.4.4
blivet-gui==2.3.0
boto3==1.39.3
botocore==1.39.3
Brlapi==0.8.4
Brotli==1.0.9
cached-property==1.5.2
cairocffi==1.3.0
CairoSVG==2.7.0
capstone==4.0.2
catfish==4.16.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
check50==3.3.11
click==8.0.4
colored-traceback==0.3.0
contourpy==1.3.2
cryptography>=42.0.0,<46.0.0
msal>=1.25.0
cssselect2==0.6.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.6
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.9.0
Django==5.2.4
django-admin-chat==0.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.2.0
docker==5.0.3
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.19.1
email_validator==2.2.0
eventlet==0.33.0
exceptiongroup==1.2.0
executing==2.2.0
fastapi==0.110.1
fb-re2==1.0.7
fedora-third-party==0.10
filelock==3.3.1
flake8==7.3.0
Flask==2.0.3
Flask-Babel==2.0.0
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.2
Flask-SocketIO==5.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.58.5
frozenlist==1.4.1
gcovr==6.0
GDAL==3.4.3
gpg==1.17.0
greenlet==1.1.2
gssapi==1.7.2
gTTS==2.5.1
h11==0.14.0
httpagentparser==1.9.2
httpcore==1.0.4
httpx==0.27.0
humanize==3.13.1
idna==3.3
iniconfig==2.1.0
initial-setup==0.3.94
intervaltree==3.1.0
ipdb==0.13.13
ipython==8.37.0
isodate==0.6.1
isort==6.0.1
itsdangerous==2.0.1
jedi==0.19.2
jellyfish==0.11.2
Jinja2==3.0.3
jmespath==1.0.0
jq==1.9.1
jsonschema==3.2.0
kaggle==1.7.4.5
kiwisolver==1.4.8
koji==1.32.0
langtable==0.0.61
ldap3==2.9.1
lib50==3.0.11
libcomps==0.1.18
libvirt-python==8.0.0
lxml==4.7.1
Mako==1.1.4
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mercurial==6.1.4
motor==3.3.1
msal-extensions==1.0.0
msrest==0.7.0
multidict==6.0.5
mutagen==1.47.0
mypy==1.16.1
mypy_extensions==1.1.0
nftables==0.1
ntlm-auth==1.5.0
numpy==2.2.6
oauthlib==3.0.2
olefile==0.46
openai==0.28.0
packaging==25.0
pandas==2.3.1
paramiko==2.12.0
parso==0.8.4
passlib==1.7.4
Paste==3.5.0
pathspec==0.10.1
pexpect==4.8.0
phonenumbers==9.0.10
phonenumberslite==9.0.10
pid==2.2.3
ping3==4.0.8
platformdirs>=2,<3
requests==2.32.3
django-filter==23.4
pdf2image>=1.16.0,<2.0.0

pluggy==1.6.0
plumbum==1.7.0
ply==3.11
portalocker==2.4.0
productmd==1.35
progressbar2==3.53.2
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==5.8.0
psycopg2>=2.9,<3.0
ptyprocess==0.6.0
pure_eval==0.2.3
pwntools==4.8.0
pwquality==1.4.4
py==1.11.0
pyasn1==0.6.1
PyAudio==0.2.11
pycairo==1.21.0
pycodestyle==2.14.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.1
pydantic==2.6.4
pydantic_core==2.16.3
pyelftools==0.28
pyenchant==3.2.2
pyflakes==3.4.0
pygame==2.5.2
pygit2==1.7.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.10.1
pykickstart==3.36
pymongo==4.5.0
PyNaCl==1.4.0
pyOpenSSL==25.1.0
pyotp==2.8.0
pyparsing==3.2.3
pyparted==3.12.0
pyrsistent==0.18.1
pyserial==3.4
PySocks==1.7.1
pystray==0.17.3
pytest==8.4.1
python-augeas==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.4.1
python-gettext==4.0
python-jose==3.5.0
python-manatools==0.0.4
python-meh==0.50
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.8.0
python-utils==2.5.6
python-xlib==0.33
pyttsx3==2.90
pytz==2023.3
pyudev==0.22.0
pywinrm==0.4.1
pyxdg==0.27
PyYAML==6.0.2
qrcode==7.3.1
redis==4.5.1
regex==2022.10.31
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
requests-ntlm==1.1.0
requests-oauthlib==2.0.0
resolvelib==0.5.5
rich==14.0.0
ROPGadget==6.7
rpm==4.17.1
rpmautospec==0.3.5
rpyc==5.0.1
rsa==4.9.1
s3transfer==0.13.0
scapy==2.5.0
selinux==3.3
sepolicy==3.3
setools==4.4.0
setproctitle==1.2.2
setroubleshoot==3.3.31
shellingham==1.5.4
simpleaudio==1.0.4
simplejson==3.17.6
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.3
soupsieve==2.3.1
SpeechRecognition==3.10.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
sshtunnel==0.4.0
stack-data==0.6.3
starlette==0.37.2
stevedore==5.5.0
submit50==3.1.5
systemd-python==234
Tempita==0.5.2
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.2
tinycss2==1.1.1
tomli==2.2.1
tox==3.28.0
tqdm==4.66.2
traitlets==5.14.3
typer==0.16.0
typing_extensions==4.10.0
tzdata==2025.2
ua-parser==0.15.0
unicorn==1.0.3
urllib3==2.2.2
user-agents==2.2.0
uvicorn==0.25.0
virtualenv==20.13.4
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.3.2
websockets==12.0
Werkzeug==2.0.3
WTForms==3.0.1
xcffib==0.11.1
xmltodict==0.12.0
yamllint==1.29.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.7.25
# LayoutLM dependencies
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
layoutlm==0.0.1
unilm==0.5.0
detectron2==0.6

# OCR and image processing
pytesseract==0.3.10
opencv-python-headless==4.6.0.66
pillow==11.3.0

# Geospatial
geojson==3.1.0
geopy==2.4.1
python-decouple==3.8
shapely>=2.0.0
djangorestframework-gis>=1.0.0

# Additional dependencies for ANDF system
psycopg2-binary>=2.9.0
django-filter>=23.0.0

